Move the `chatReady` and `whispersReady` socket events to an HTTP request that happens on load instead.
This also simplifies a little bit the way we're sending the joined
channels data (previously we had an unnecessary DB call).

Also, this refactors the whisper service a little bit so that `init`
data is not sent when the user connects. Now it's only sent when a new
whisper session is started. This is more inline with how chat service
works as well, since we had a bunch of issues with channels ordering due
to that there.